requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
clarifai==11.2.3
clarifai-grpc==11.3.2
clarifai-protocol==0.0.22
click==8.1.8
cohere==5.15.0
colorama==0.4.6
configparser==3.8.1
contextlib2==21.6.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
Django==5.1.7
django-allauth==65.7.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-storages==1.14.6
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
fastavro==1.10.0
filelock==3.18.0
firebase-admin==6.7.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gast==0.6.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-cloud-vision==3.10.1
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gotrue==2.12.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonschema==2.6.0
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.2
namex==0.0.8
networkx==3.4.2
numpy==2.1.1
oauthlib==3.2.2
openai==1.70.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
postgrest==1.0.1
proglog==0.1.11
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytesseract==0.3.13
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-rapidjson==1.20
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
schema==0.7.5
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
SpeechRecognition==3.14.2
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
termcolor==3.0.1
tf_keras==2.19.0
thop==0.1.1.post2209072238
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
tritonclient==2.34.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.106
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.9.0
wrapt==1.17.2
yarl==1.19.0
youtube-transcript-api==1.0.3
